Initialization!
(copy/pasted this from the program, realized it would be a better idea
to put it here then to have my giant mess/wall of text in the program)

This program is going to convert between DNA, RNA, and amino acid
sequences!

#PSUEDOCODE YEAH
#DNA -> RNA
#RNA -> Amino Acids
#Amino Acids -> Protein (vague folding simulation)
#
#Okay so basically
'''
Converting to and from DNA and RNA will be pretty easy.
Only trouble I forsee is implementing a system that
allows you to choose with strand of the DNA you wish
to use, and then convert that to a single strand of RNA.

It doesn't matter because you don't lose any information,
but biology is complicated and I don't know if the strand
chosen matters or is important to some processes.

Going from RNA to an Amino Acid chain will be pretty easy
as well, but will require some additional research on my
part because I don't know what base pairs make what amino
acids. It can't be that hard though. There's only like
20 of them or so, right?

As for proteins, I have no fucking clue how they fold.
I know it's complicated and looks pretty scary
(helices and fold and whatnot!!), but maybe I could
detect patterns in the amino acid placement and suggest
a very rudimentary folding path?? It would be really simple
and probably mostly scientifically incorrect. This idea
probably isn't going to get implemented.
'''